@@ -16,7 +16,7 @@ perceptor = "pcr.perceptor:main"
 pcr = "pcr.perceptor:main"
 
 [tool.poetry.dependencies]
-python = ">=3.10,<3.13"
+python = ">=3.11,<3.13"
 rich = "^13.7.1"
 lief = "^0.14.1"
 keystone-engine = "^0.9.2"
@@ -30,6 +30,8 @@ xmltodict = "^0.13.0"
 peewee = "^3.16.3"
 faker = "^19.13.0"
 pythonnet = "^3.0.2"
+coffipy = {git = "https://github.com/Veids/COFFI", rev = "coffipy"}
+furikuripy = {git = "https://github.com/Veids/furikuripy"}
 
 [tool.poetry.group.dev.dependencies]
 ipython = "^8.14.0"
